pyproject.toml
[tool.poetry]
name = "megumin-bot"
version = "0.1.0"
description = ""
authors = ["glitchy <thepatheticweebgamer@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
naff = { extras = ["all"], version = "^1.2.0" }
rich = "^12.4.4"
pygelbooru = "^0.5.0"
NHentai-API = "^0.0.19"
asyncpraw = "^7.5.0"
pet-pet-gif = "^1.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
